requirements/requirements.txt
fastapi
SQLAlchemy
python-multipart
uvicorn
python-jose
bcrypt
passlib[bcrypt]